refactor: Split message and topic classes into own modules
