+++ aiomqtt/exceptions.py
